refactor(ui): 파일 분리 구조 삭제, 타입과 상수 순서 수정, 오타 수정
- 두개의 파일의 순환 참조로 인해 일부 번들러 문제 발생 가능성이 생겨 분리된 파일 하나로 합침
- 변수를 참조하여 타입을 구현하여 타입과 변수 순서 변경
